# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TECA API",
    "description": "\nAPI del e-commerce de muebles **TECA — Para toda la vida**.\n\nCubre la tienda pública (catálogo, carrito, checkout de invitado o con cuenta),\nel área del cliente (pedidos, direcciones, métodos de pago, devoluciones) y el\npanel de administración (usuarios internos, productos, pedidos, cupones,\nauditoría y finanzas).\n\n### Autenticación\nTodas las rutas protegidas usan **JWT Bearer**:\n\n1. `POST /auth/login` con correo y contraseña.\n2. Copia el `access_token` de la respuesta.\n3. Pulsa **Authorize** (arriba a la derecha) y pégalo.\n\n### Roles\n| Rol | Alcance |\n|---|---|\n| `cliente` | Tienda: carrito, pedidos propios, cuenta, reseñas |\n| `admin` | Acceso total al panel |\n| `editor` | Productos |\n| `encargado` | Productos, pedidos y devoluciones |\n| `vendedor` | Pedidos (y lectura de productos) |\n| `soporte` | Devoluciones y lectura de pedidos |\n| `finanzas` | Reportes financieros |\n\n### Convenciones\n- Moneda: balboas (B/.), equivalente a USD.\n- Los identificadores son `ObjectId` de Mongo expuestos como `id` (string).\n- Errores en formato `{\"detail\": \"mensaje\"}` con el código HTTP correspondiente.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/orders/checkout": {
      "post": {
        "tags": [
          "Pedidos"
        ],
        "summary": "Crear un pedido (checkout)",
        "description": "Cierra la compra y genera el pedido. Es el paso *Pago → Confirmación* del checkout.\n\n**Funciona con y sin cuenta.** El token es opcional: si se envía, el pedido queda asociado al usuario y se le vacía el carrito; si no, se registra como pedido de **invitado** (por eso `email` es obligatorio siempre — es el correo con el que luego podrá consultar su pedido).\n\n**Qué hace, en orden:**\n1. Valida que cada producto exista, esté activo y tenga stock suficiente.\n2. Recalcula precios **en el servidor** (nunca confía en los totales del frontend) y aplica el cupón si se envió.\n3. Genera el número correlativo del pedido (`TEC-2026-001`) con un contador atómico por año.\n4. Descuenta el stock de cada producto.\n5. Deja el pedido en estado `confirmado` con su primer registro en la línea de tiempo.\n\n**Métodos de pago** (`payment_method`): `card`, `paypal`, `bitcoin`, `yappy`. El cobro está **simulado**: el pedido nace con `payment_status: pagado`; aquí se integraría la pasarela real.\n\n**Errores:** `404` producto inexistente · `409` stock insuficiente · `400` cupón inválido.",
        "operationId": "checkout_orders_checkout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Datos del checkout"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Producto no encontrado"
          },
          "409": {
            "description": "Stock insuficiente para algún producto"
          },
          "400": {
            "description": "Cupón inválido o expirado"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error de validación"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/orders/mine": {
      "get": {
        "tags": [
          "Pedidos"
        ],
        "summary": "Mis pedidos",
        "description": "Pantalla **Mis pedidos** del área de cliente. Devuelve todos los pedidos del usuario autenticado, del más reciente al más antiguo.\n\nUn arreglo vacío es la señal para mostrar el estado *«Todavía no tienes pedidos»*.\n\n🔒 Requiere token.",
        "operationId": "my_orders_orders_mine_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/orders/lookup": {
      "post": {
        "tags": [
          "Pedidos"
        ],
        "summary": "Consultar un pedido sin cuenta (invitado)",
        "description": "Permite a quien compró como invitado rastrear su pedido: *«Puedes consultar el estado de tu pedido usando el número de pedido y tu correo electrónico»*.\n\nExige **ambos datos** (`order_number` + `email`) y deben coincidir, de modo que un número de pedido adivinado no basta para ver los datos de otra persona.\n\n**No requiere token. Errores:** `404` si la combinación no coincide.",
        "operationId": "lookup_order_orders_lookup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Consulta de pedido (invitado)"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Pedido no encontrado"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error de validación"
                }
              }
            }
          }
        }
      }
    },
    "/orders/{order_number}": {
      "get": {
        "tags": [
          "Pedidos"
        ],
        "summary": "Detalle y seguimiento de un pedido",
        "description": "Pantalla **Pedido TEC-2026-001**: información de envío, productos, resumen de totales y `tracking_number` (número de guía del courier).\n\nEl campo `status_history` trae cada estado con su fecha y hora, que es lo que dibuja la línea de tiempo *Pedido confirmado → En preparación → En camino → Entregado*.\n\n**Visibilidad:** un `cliente` solo puede ver sus propios pedidos; los roles internos pueden consultar cualquiera.\n\n🔒 Requiere token. **Errores:** `404` si no existe o no te pertenece.",
        "operationId": "get_order_orders__order_number__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Número del pedido. Ej. «TEC-2026-001»",
              "title": "Order Number"
            },
            "description": "Número del pedido. Ej. «TEC-2026-001»"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Pedido no encontrado"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error de validación"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}